refactor: auto-split test_ball_flight_physics.py
