+++ tests/unit/ball_flight_physics/test_edge_cases.py
@@ -1,0 +1,192 @@
+"""Unit tests for ball flight physics simulation.
+
+Tests cover:
+- Physical properties validation
+- Launch conditions and environmental effects
+- Trajectory simulation correctness
+- Force calculations (gravity, drag, Magnus)
+- Output metrics accuracy
+"""
+
+import math
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.physics.rust_kernel import is_rust_available
+
+pytestmark = pytest.mark.skipif(
+    not is_rust_available(),
+    reason="upstream-physics Rust kernel not installed (pip install upstream-drift[rust])",
+)
+
+from src.shared.python.physics.ball_flight_physics import (  # noqa: E402
+    BallFlightSimulator,
+    BallProperties,
+    EnvironmentalConditions,
+    LaunchConditions,
+    TrajectoryPoint,
+)
+
+# =============================================================================
+# Test Fixtures
+# =============================================================================
+
+
+@pytest.fixture
+def default_ball() -> BallProperties:
+    """Default regulation golf ball properties."""
+    return BallProperties()
+
+
+@pytest.fixture
+def default_environment() -> EnvironmentalConditions:
+    """Default sea-level, no wind conditions."""
+    return EnvironmentalConditions()
+
+
+@pytest.fixture
+def simulator() -> BallFlightSimulator:
+    """Default simulator with regulation ball and standard conditions."""
+    return BallFlightSimulator()
+
+
+@pytest.fixture
+def driver_launch() -> LaunchConditions:
+    """Typical driver launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~73 m/s (163 mph)
+    - Launch angle: ~11 degrees
+    - Backspin: ~2500 rpm
+    """
+    return LaunchConditions(
+        velocity=73.0,
+        launch_angle=math.radians(11.0),
+        spin_rate=2500.0,
+    )
+
+
+@pytest.fixture
+def iron_7_launch() -> LaunchConditions:
+    """Typical 7-iron launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~53 m/s (118 mph)
+    - Launch angle: ~16 degrees
+    - Backspin: ~7000 rpm
+    """
+    return LaunchConditions(
+        velocity=53.0,
+        launch_angle=math.radians(16.0),
+        spin_rate=7000.0,
+    )
+
+
+# =============================================================================
+# BallProperties Tests
+# =============================================================================
+
+
+# =============================================================================
+# LaunchConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# EnvironmentalConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# BallFlightSimulator Initialization Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Simulation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Physics Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Force Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Metric Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Real-World Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Edge Cases and Error Handling
+# =============================================================================
+
+
+class TestEdgeCases:
+    """Tests for edge cases and boundary conditions."""
+
+    def test_very_low_velocity(self, simulator: BallFlightSimulator) -> None:
+        """Test trajectory with very low velocity."""
+        low_velocity_launch = LaunchConditions(
+            velocity=5.0, launch_angle=math.radians(45.0), spin_rate=0.0
+        )
+        trajectory = simulator.simulate_trajectory(low_velocity_launch, max_time=2.0)
+
+        assert len(trajectory) > 0
+        carry = simulator.calculate_carry_distance(trajectory)
+        assert carry > 0  # Should still move forward
+
+    def test_very_high_spin(self, simulator: BallFlightSimulator) -> None:
+        """Test trajectory with extreme spin rate."""
+        high_spin_launch = LaunchConditions(
+            velocity=50.0, launch_angle=math.radians(15.0), spin_rate=10000.0
+        )
+        trajectory = simulator.simulate_trajectory(high_spin_launch, max_time=8.0)
+
+        # Should still produce valid trajectory
+        assert len(trajectory) > 0
+        max_height = simulator.calculate_max_height(trajectory)
+        assert max_height > 0
+
+    def test_zero_launch_angle(self, simulator: BallFlightSimulator) -> None:
+        """Test horizontal launch (zero launch angle).
+
+        When launched horizontally from ground level (z=0) with no spin,
+        the ball immediately contacts the ground due to gravity.
+        """
+        horizontal_launch = LaunchConditions(
+            velocity=50.0, launch_angle=0.0, spin_rate=0.0
+        )
+        trajectory = simulator.simulate_trajectory(horizontal_launch, max_time=3.0)
+
+        # Should have at least one point (the initial state)
+        assert len(trajectory) >= 1
+        # Ball should not rise above starting height
+        max_height = max(p.position[2] for p in trajectory)
+        assert max_height <= 0.1  # Within 10cm of ground (accounting for numerics)
+
+
+# =============================================================================
+# Wind Effect Tests
+# =============================================================================
+
+
+# =============================================================================
+# Spin Decay Tests
+# =============================================================================
