+++ tests/unit/ball_flight_physics/test_launch_conditions.py
@@ -1,0 +1,177 @@
+"""Unit tests for ball flight physics simulation.
+
+Tests cover:
+- Physical properties validation
+- Launch conditions and environmental effects
+- Trajectory simulation correctness
+- Force calculations (gravity, drag, Magnus)
+- Output metrics accuracy
+"""
+
+import math
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.physics.rust_kernel import is_rust_available
+
+pytestmark = pytest.mark.skipif(
+    not is_rust_available(),
+    reason="upstream-physics Rust kernel not installed (pip install upstream-drift[rust])",
+)
+
+from src.shared.python.physics.ball_flight_physics import (  # noqa: E402
+    BallFlightSimulator,
+    BallProperties,
+    EnvironmentalConditions,
+    LaunchConditions,
+    TrajectoryPoint,
+)
+
+# =============================================================================
+# Test Fixtures
+# =============================================================================
+
+
+@pytest.fixture
+def default_ball() -> BallProperties:
+    """Default regulation golf ball properties."""
+    return BallProperties()
+
+
+@pytest.fixture
+def default_environment() -> EnvironmentalConditions:
+    """Default sea-level, no wind conditions."""
+    return EnvironmentalConditions()
+
+
+@pytest.fixture
+def simulator() -> BallFlightSimulator:
+    """Default simulator with regulation ball and standard conditions."""
+    return BallFlightSimulator()
+
+
+@pytest.fixture
+def driver_launch() -> LaunchConditions:
+    """Typical driver launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~73 m/s (163 mph)
+    - Launch angle: ~11 degrees
+    - Backspin: ~2500 rpm
+    """
+    return LaunchConditions(
+        velocity=73.0,
+        launch_angle=math.radians(11.0),
+        spin_rate=2500.0,
+    )
+
+
+@pytest.fixture
+def iron_7_launch() -> LaunchConditions:
+    """Typical 7-iron launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~53 m/s (118 mph)
+    - Launch angle: ~16 degrees
+    - Backspin: ~7000 rpm
+    """
+    return LaunchConditions(
+        velocity=53.0,
+        launch_angle=math.radians(16.0),
+        spin_rate=7000.0,
+    )
+
+
+# =============================================================================
+# BallProperties Tests
+# =============================================================================
+
+
+# =============================================================================
+# LaunchConditions Tests
+# =============================================================================
+
+
+class TestLaunchConditions:
+    """Tests for LaunchConditions dataclass."""
+
+    def test_default_spin_axis(self) -> None:
+        """Test that default spin axis produces upward lift for backspin.
+
+        Default axis is [0, -1, 0] which when crossed with forward velocity
+        produces upward lift (Magnus force in +z direction).
+        """
+        launch = LaunchConditions(velocity=50.0, launch_angle=0.1)
+        expected_axis = np.array([0.0, -1.0, 0.0])
+        assert launch.spin_axis is not None
+        np.testing.assert_array_almost_equal(launch.spin_axis, expected_axis)
+
+    def test_custom_spin_axis(self) -> None:
+        """Test custom spin axis is preserved."""
+        custom_axis = np.array([0.5, 0.5, 0.0])
+        launch = LaunchConditions(
+            velocity=50.0, launch_angle=0.1, spin_axis=custom_axis
+        )
+        assert launch.spin_axis is not None
+        np.testing.assert_array_almost_equal(launch.spin_axis, custom_axis)
+
+    def test_zero_spin_rate_allowed(self) -> None:
+        """Test that zero spin rate is valid."""
+        launch = LaunchConditions(velocity=50.0, launch_angle=0.1, spin_rate=0.0)
+        assert launch.spin_rate == 0.0
+
+
+# =============================================================================
+# EnvironmentalConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# BallFlightSimulator Initialization Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Simulation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Physics Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Force Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Metric Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Real-World Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Edge Cases and Error Handling
+# =============================================================================
+
+
+# =============================================================================
+# Wind Effect Tests
+# =============================================================================
+
+
+# =============================================================================
+# Spin Decay Tests
+# =============================================================================
