+++ tests/unit/ball_flight_physics/test_real_world_validation.py
@@ -1,0 +1,216 @@
+"""Unit tests for ball flight physics simulation.
+
+Tests cover:
+- Physical properties validation
+- Launch conditions and environmental effects
+- Trajectory simulation correctness
+- Force calculations (gravity, drag, Magnus)
+- Output metrics accuracy
+"""
+
+import math
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.physics.rust_kernel import is_rust_available
+
+pytestmark = pytest.mark.skipif(
+    not is_rust_available(),
+    reason="upstream-physics Rust kernel not installed (pip install upstream-drift[rust])",
+)
+
+from src.shared.python.physics.ball_flight_physics import (  # noqa: E402
+    BallFlightSimulator,
+    BallProperties,
+    EnvironmentalConditions,
+    LaunchConditions,
+    TrajectoryPoint,
+)
+
+# =============================================================================
+# Test Fixtures
+# =============================================================================
+
+
+@pytest.fixture
+def default_ball() -> BallProperties:
+    """Default regulation golf ball properties."""
+    return BallProperties()
+
+
+@pytest.fixture
+def default_environment() -> EnvironmentalConditions:
+    """Default sea-level, no wind conditions."""
+    return EnvironmentalConditions()
+
+
+@pytest.fixture
+def simulator() -> BallFlightSimulator:
+    """Default simulator with regulation ball and standard conditions."""
+    return BallFlightSimulator()
+
+
+@pytest.fixture
+def driver_launch() -> LaunchConditions:
+    """Typical driver launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~73 m/s (163 mph)
+    - Launch angle: ~11 degrees
+    - Backspin: ~2500 rpm
+    """
+    return LaunchConditions(
+        velocity=73.0,
+        launch_angle=math.radians(11.0),
+        spin_rate=2500.0,
+    )
+
+
+@pytest.fixture
+def iron_7_launch() -> LaunchConditions:
+    """Typical 7-iron launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~53 m/s (118 mph)
+    - Launch angle: ~16 degrees
+    - Backspin: ~7000 rpm
+    """
+    return LaunchConditions(
+        velocity=53.0,
+        launch_angle=math.radians(16.0),
+        spin_rate=7000.0,
+    )
+
+
+# =============================================================================
+# BallProperties Tests
+# =============================================================================
+
+
+# =============================================================================
+# LaunchConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# EnvironmentalConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# BallFlightSimulator Initialization Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Simulation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Physics Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Force Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Metric Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Real-World Validation Tests
+# =============================================================================
+
+
+class TestRealWorldValidation:
+    """Tests validating against known real-world data.
+
+    Reference data from TrackMan/FlightScope:
+    - Driver: 163 mph ball speed, 11° launch, 2500 rpm → ~275 yards carry
+    - 7-iron: 118 mph ball speed, 16° launch, 7000 rpm → ~160 yards carry
+    """
+
+    def test_driver_carry_reasonable_range(
+        self, simulator: BallFlightSimulator, driver_launch: LaunchConditions
+    ) -> None:
+        """Test driver carry distance is in reasonable range.
+
+        The current empirical model targets:
+        - Driver (73 m/s, 11°, 2500 rpm): ~200 yards
+
+        Note: Model accuracy depends on lift/drag coefficient tuning.
+        Wider tolerance allows for different coefficient sets.
+        """
+        trajectory = simulator.simulate_trajectory(driver_launch, max_time=8.0)
+        carry_m = simulator.calculate_carry_distance(trajectory)
+        carry_yards = carry_m * 1.09361  # Convert to yards
+
+        # Broad tolerance: 150-350 yards to accommodate model variations
+        assert 150 < carry_yards < 350, f"Carry was {carry_yards:.1f} yards"
+
+    def test_iron_7_carry_reasonable_range(
+        self, simulator: BallFlightSimulator, iron_7_launch: LaunchConditions
+    ) -> None:
+        """Test 7-iron carry distance is in reasonable range.
+
+        The current empirical model targets:
+        - 7-iron (53 m/s, 16°, 7000 rpm): ~165 yards
+        """
+        trajectory = simulator.simulate_trajectory(iron_7_launch, max_time=8.0)
+        carry_m = simulator.calculate_carry_distance(trajectory)
+        carry_yards = carry_m * 1.09361
+
+        # Should be between 100 and 220 yards
+        assert 100 < carry_yards < 220, f"Carry was {carry_yards:.1f} yards"
+
+    def test_driver_max_height_reasonable(
+        self, simulator: BallFlightSimulator, driver_launch: LaunchConditions
+    ) -> None:
+        """Test driver max height is reasonable.
+
+        Model produces heights in range 10-25m for driver trajectory.
+        """
+        trajectory = simulator.simulate_trajectory(driver_launch, max_time=8.0)
+        max_height = simulator.calculate_max_height(trajectory)
+
+        # Allow 5-50m range for model variations
+        assert 5 < max_height < 50, f"Max height was {max_height:.1f}m"
+
+    def test_driver_flight_time_reasonable(
+        self, simulator: BallFlightSimulator, driver_launch: LaunchConditions
+    ) -> None:
+        """Test driver flight time is reasonable.
+
+        Model produces flight times of 3-6 seconds for driver.
+        """
+        trajectory = simulator.simulate_trajectory(driver_launch, max_time=10.0)
+        flight_time = simulator.calculate_flight_time(trajectory)
+
+        # Allow 2-8 seconds for model variations
+        assert 2 < flight_time < 8, f"Flight time was {flight_time:.1f}s"
+
+
+# =============================================================================
+# Edge Cases and Error Handling
+# =============================================================================
+
+
+# =============================================================================
+# Wind Effect Tests
+# =============================================================================
+
+
+# =============================================================================
+# Spin Decay Tests
+# =============================================================================
