+++ tests/unit/ball_flight_physics/test_simulator_initialization.py
@@ -1,0 +1,170 @@
+"""Unit tests for ball flight physics simulation.
+
+Tests cover:
+- Physical properties validation
+- Launch conditions and environmental effects
+- Trajectory simulation correctness
+- Force calculations (gravity, drag, Magnus)
+- Output metrics accuracy
+"""
+
+import math
+from unittest.mock import MagicMock
+
+import numpy as np
+import pytest
+from src.shared.python.physics.rust_kernel import is_rust_available
+
+pytestmark = pytest.mark.skipif(
+    not is_rust_available(),
+    reason="upstream-physics Rust kernel not installed (pip install upstream-drift[rust])",
+)
+
+from src.shared.python.physics.ball_flight_physics import (  # noqa: E402
+    BallFlightSimulator,
+    BallProperties,
+    EnvironmentalConditions,
+    LaunchConditions,
+    TrajectoryPoint,
+)
+
+# =============================================================================
+# Test Fixtures
+# =============================================================================
+
+
+@pytest.fixture
+def default_ball() -> BallProperties:
+    """Default regulation golf ball properties."""
+    return BallProperties()
+
+
+@pytest.fixture
+def default_environment() -> EnvironmentalConditions:
+    """Default sea-level, no wind conditions."""
+    return EnvironmentalConditions()
+
+
+@pytest.fixture
+def simulator() -> BallFlightSimulator:
+    """Default simulator with regulation ball and standard conditions."""
+    return BallFlightSimulator()
+
+
+@pytest.fixture
+def driver_launch() -> LaunchConditions:
+    """Typical driver launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~73 m/s (163 mph)
+    - Launch angle: ~11 degrees
+    - Backspin: ~2500 rpm
+    """
+    return LaunchConditions(
+        velocity=73.0,
+        launch_angle=math.radians(11.0),
+        spin_rate=2500.0,
+    )
+
+
+@pytest.fixture
+def iron_7_launch() -> LaunchConditions:
+    """Typical 7-iron launch conditions.
+
+    Based on PGA Tour average:
+    - Ball speed: ~53 m/s (118 mph)
+    - Launch angle: ~16 degrees
+    - Backspin: ~7000 rpm
+    """
+    return LaunchConditions(
+        velocity=53.0,
+        launch_angle=math.radians(16.0),
+        spin_rate=7000.0,
+    )
+
+
+# =============================================================================
+# BallProperties Tests
+# =============================================================================
+
+
+# =============================================================================
+# LaunchConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# EnvironmentalConditions Tests
+# =============================================================================
+
+
+# =============================================================================
+# BallFlightSimulator Initialization Tests
+# =============================================================================
+
+
+class TestSimulatorInitialization:
+    """Tests for BallFlightSimulator initialization."""
+
+    def test_ball_flight_physics_default_initialization(self) -> None:
+        """Test simulator initializes with default components."""
+        sim = BallFlightSimulator()
+        assert sim.ball is not None
+        assert sim.environment is not None
+
+    def test_custom_ball(self) -> None:
+        """Test simulator with custom ball."""
+        custom_ball = BallProperties(mass=0.05)
+        sim = BallFlightSimulator(ball=custom_ball)
+        assert sim.ball.mass == 0.05
+
+    def test_custom_environment(self) -> None:
+        """Test simulator with custom environment."""
+        custom_env = EnvironmentalConditions(altitude=1500.0)
+        sim = BallFlightSimulator(environment=custom_env)
+        assert sim.environment.altitude == 1500.0
+
+
+# =============================================================================
+# Trajectory Simulation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Physics Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Force Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Metric Calculation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Trajectory Analysis Tests
+# =============================================================================
+
+
+# =============================================================================
+# Real-World Validation Tests
+# =============================================================================
+
+
+# =============================================================================
+# Edge Cases and Error Handling
+# =============================================================================
+
+
+# =============================================================================
+# Wind Effect Tests
+# =============================================================================
+
+
+# =============================================================================
+# Spin Decay Tests
+# =============================================================================
